+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -15,6 +15,7 @@ group: Deployment & Configs
 - Added 'Add Drafting View' component.
 - Added 'Element Face Paint' component.
 - Added 'Document Warnings' component.
+- Added 'Built-In Failure Definitions' picker.
 
 ### RC